Adding first cut at room rescheduling rules

+++ scratch/dax/gaia_u/main.cpp
@@ -34,7 +34,7 @@ int main(int argc, const char**argv) {
         show_all();
     }
     else
-    if (argc == 2)
+    if (argc == 2 && (strcmp(argv[1], "-d") == 0))
     {
         delete_all();
     }
